fix(lib): resolve lint errors in packages/lib
- Fix unicorn/no-array-callback-reference in command-runner.ts using pipe()
- Fix no-restricted-syntax push-spread in create-project.ts
- Extract docker identity conflict logic to docker-identity.ts (max-lines)
